User update implemented, needs further testing
We are making some changes. Here, the new JWT token is generated based on the user id. When user id is decoded, we are trying to match the user using not only the JWT token, but also their id as well. Thus making sure that a user cant change other users information. The goal is to let the user to change their information only if he has a valid token as well as his own object id.